+++ StripePaymentSheet/StripePaymentSheet/Source/Helpers/Images.swift
@@ -53,6 +53,7 @@ enum Image: String, CaseIterable, ImageMaker {
     case pm_type_touch_n_go = "icon-pm-touch_n_go"
     case pm_type_truemoney = "icon-pm-truemoney"
     case pm_type_shopeepay = "icon-pm-shopeepay"
+    case pm_type_ng_wallet = "icon-pm-ng_wallet"
     case pm_type_gopay = "icon-pm-gopay"
     case pm_type_grabpay = "icon-pm-grabpay"
     case pm_type_ideal = "icon-pm-ideal"