+++ CHANGELOG.md
@@ -10,6 +10,7 @@ MINOR
 * [Added] Added `additionalHeaders` support to the `STPAPIClient` GET, POST, and DELETE APIs exposed through the `STP` SPI.
 
 ### PaymentSheet
+* [Added] Added Naira Wallet API bindings and support for payments in PaymentSheet.
 * [Added] Added ShopeePay API bindings and support for payments in PaymentSheet.
 * [Added] Added TrueMoney API bindings and support for payments and setup in PaymentSheet.
 * [Added] Added Touch 'n Go API bindings and support for payments and setup in PaymentSheet.